feat: migrate compute_port_geometry
