+++ .env.template
@@ -19,9 +19,13 @@ LOGGING_LEVEL=INFO
 
 AMP_DATA_CSV_PATH=
 PORT_DATA_CSV_PATH=
+PORT_POLYGON_DATA_CSV_PATH=
 VESSEL_DATA_CSV_PATH=
 
 VESSEL_POSITIONS_DATA_CSV_PATH=
+
+#PORT_RADIUS_M=3000
+#PORT_RESOLUTION=10
 
 
 ###############################################################################################
